% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-bounds.R
\name{median-bounds}
\alias{median-bounds}
\alias{median_bounds}
\alias{median_bounds.default}
\title{Bounds on an unknown median}
\usage{
median_bounds(x, na.rm.amount = 0, even = c("mean", "low", "high"), nna = NULL)

\method{median_bounds}{default}(x, na.rm.amount = 0, even = c("mean", "low", "high"), nna = NULL)
}
\arguments{
\item{x}{Vector that can be ordered using \code{\link[=sort]{sort()}}. It will be searched for
its possible medians.}

\item{na.rm.amount, even}{Passed on to \code{\link[=median2]{median2()}}.}

\item{nna}{Integer. Ignore unless the function is used as a helper. It is
like \code{needs_prep} in \code{\link[=median_count_tolerable]{median_count_tolerable()}} except it can submit the
number of missing values.}
}
\value{
Vector of length 2 (never \code{NA}). Its type is double if \code{x} is numeric
(double or integer) or if \code{c(-Inf, Inf)} is returned, which is always
double. Otherwise, it has the same type as \code{x}.
}
\description{
\code{median_bounds()} computes the minimal and maximal possible
median values. This is helpful if \code{\link[=median2]{median2()}} returns \code{NA}: the median
can't be determined, but at least it might have lower and upper bounds.

It is used within \code{\link[=median_table]{median_table()}} to compute the \code{lower} and \code{upper}
columns.
}
\details{
Two edge cases may occur:
\itemize{
\item If the median can be precisely determined, \code{median_bounds()} returns two
identical values. This is the same case in which \code{\link[=median2]{median2()}} would return
a non-\code{NA} value, and \code{\link[=median_plot_errorbar]{median_plot_errorbar()}} would show a "ring of
certainty".
\item If the number of missing values is so high that a continuous array of
them could extend from the start or end of \code{x} into the median position,
the data do not constrict the median to fall in between any finite bounds.
The function will then return \code{c(-Inf, Inf)} because such an array of \code{NA}s
would act as a tunnel to negative or positive infinity, enabling the median
to assume indefinitely low or high values.
}

Like \code{\link[=median2]{median2()}}, this function is generic, so methods can be defined for
other classes. This documentation describes the default method.
}
\examples{
# Lower and upper bounds can be found,
# even though the precise median is unknown:
median_bounds(c(7, 7, 8, 9, NA))
median_bounds(c(7, 7, 7, 8, 9, 9, NA, NA))

# Too many missing values, so there is no finite range:
median_bounds(c(7, 7, 8, 9, NA, NA, NA, NA))
}
\author{
Lukas Jung, R Core Team
}
