% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-possible.R
\name{median-possible}
\alias{median-possible}
\alias{median_possible_values}
\alias{median_range}
\alias{median_possible_values.default}
\alias{median_range.default}
\title{Possible medians}
\usage{
median_possible_values(x)

median_range(x)

\method{median_possible_values}{default}(x)

\method{median_range}{default}(x)
}
\arguments{
\item{x}{Numeric or similar. Vector to search for its possible medians.}
}
\value{
Vector of the same type as \code{x}. Its values are unique and sorted. It
always has length 2 when calling \code{median_range()}.

If the median can be determined (i.e., \code{median2()} would return a non-\code{NA}
value), \code{median_possible_values()} returns just one value, and
\code{median_range()} returns two identical values.

If there are missing values at the median position, the set of possible
medians cannot be determined. The functions will then return \code{NA} or \code{c(NA, NA)}, respectively.
}
\description{
These functions are helpful if \code{median2()} returns \code{NA}:
\itemize{
\item \code{median_possible_values()} returns all values that might possibly be the
median, given the number of missing values.
\item \code{median_range()} returns the minimal and maximal possible median values.
}
}
\details{
Like \code{median2()}, these functions are generic, so methods can be
defined for other classes. This documentation describes the default
methods.
}
\examples{
# All possible values,
# plus the range:
x1 <- c(7, 7, 7, 8, 9, 9, NA, NA)
median_possible_values(x1)
median_range(x1)

# The output of both functions
# is often the same:
x2 <- c(7, 7, 7, 8, NA, NA)
median_possible_values(x2)
median_range(x2)

# If there are missing values at
# the median, even these functions
# return `NA`:
x3 <- c(1, 1, 2, NA, NA, NA)
median_possible_values(x3)
median_range(x3)
}
\author{
Lukas Jung, R Core Team
}
