% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-range.R
\name{median-range}
\alias{median-range}
\alias{median_range}
\alias{median_range.default}
\title{Bounds on an unknown median}
\usage{
median_range(x, na.rm.amount = 0, even = c("mean", "low", "high"))

\method{median_range}{default}(x, na.rm.amount = 0, even = c("mean", "low", "high"))
}
\arguments{
\item{x}{Numeric or similar. Vector to search for its possible medians.}

\item{na.rm.amount, even}{Passed on to \code{\link[=median2]{median2()}}.}
}
\value{
Vector of length 2 and the same type as \code{x}.

If the median can be determined (i.e., \code{\link[=median2]{median2()}} would return a
non-\code{NA} value), \code{median_range()} returns two identical values.

If there could be missing values at the median position, the median range
cannot be determined. The function will then return \code{c(NA, NA)} but of the
same type as \code{x}.
}
\description{
\code{median_range()} computes the minimal and maximal possible
median values. This is helpful if \code{\link[=median2]{median2()}} returns \code{NA}: the median
can't be determined, but at least its bounds might be known.

It is used within \code{\link[=median_table]{median_table()}} to compute the \code{min} and \code{max}
columns.
}
\details{
Like \code{\link[=median2]{median2()}}, this function is generic, so methods can be
defined for other classes. This documentation describes the default method.
}
\examples{
# Lower and upper bounds are known,
# even though `median2()` would return `NA`:
median_range(c(7, 7, 8, 9, NA))
median_range(c(7, 7, 7, 8, 9, 9, NA, NA))

# Too many missing values, so the range is unknown:
median_range(c(7, 7, 8, 9, NA, NA, NA, NA))
}
\author{
Lukas Jung, R Core Team
}
