% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median2.R
\name{median2}
\alias{median2}
\alias{median2.default}
\title{Drop-in replacement for \code{median()}}
\usage{
median2(
  x,
  na.rm = FALSE,
  na.rm.amount = 0,
  na.rm.from = c("first", "last", "random"),
  even = c("mean", "low", "high"),
  ...
)

\method{median2}{default}(
  x,
  na.rm = FALSE,
  na.rm.amount = 0,
  na.rm.from = c("first", "last", "random"),
  even = c("mean", "low", "high"),
  ...
)
}
\arguments{
\item{x}{Vector that can be ordered using \code{\link[=sort]{sort()}}. It will be searched for
its median.}

\item{na.rm}{Logical. If set to \code{TRUE}, missing values are removed before
computation proceeds. Default is \code{FALSE}.}

\item{na.rm.amount}{Numeric. Alternative to \code{na.rm} that only removes a
specified number of missing values. Default is \code{0}.}

\item{na.rm.from}{String. If \code{na.rm.amount} is used, from which position in
\code{x} should missing values be removed? Options are \code{"first"}, \code{"last"}, and
\code{"random"}. Default is \code{"first"}.}

\item{even}{String. What to return if \code{x} has an even length and contains no
missing values (or they were removed). The default, \code{"mean"}, averages the
two central values of the sorted vector, \code{"low"} returns the lower central
value, and \code{"high"} returns the higher one. Note that \code{"mean"} is only
allowed is \code{x} is numeric.}

\item{...}{Optional further arguments for methods. Not used in the default
method.}
}
\value{
Length-1 vector of type double if the input is numeric, and the same
type as \code{x} otherwise. This is tested by \code{\link[=is.numeric]{is.numeric()}}, so factors and
dates do not count as numeric.
}
\description{
\code{median2()} computes the sample median. By default, it works
like \code{\link[=median]{median()}} from base R, with these exceptions:
\itemize{
\item If one or more values are missing, \code{median2()} checks if the median can
be determined nevertheless. \code{median()} always returns \code{NA} in this case,
but \code{median2()} only returns \code{NA} if the median is genuinely unknown.
\item Non-numeric data, including dates and factors, require one of
\code{even = "low"} and \code{even = "high"}. This option doesn't exist in
\code{median()}. It avoids "computing the mean" of the two central values of
sorted vectors with an even length when no such operation is possible,
e.g., with strings.
\item The return type is always double if the input vector is numeric (i.e.,
double or integer). This is consistent and predictable, regardless of the
length being even or odd.
}
}
\details{
\code{median2()} is a generic function, so new methods can be defined for
it. As with \code{\link[stats:median]{stats::median()}} from base R, the default method described
here should work for most classes for which a median is a reasonable
concept (e.g., "\code{\link{Date}}").

If a new method is necessary, please make sure it deals with missing values
like the default method does. See
\href{https://lhdjung.github.io/naidem/articles/algorithm.html}{\emph{Implementing
the algorithm}} for further details.
}
\examples{
# If no values are missing,
# it works mostly like `median()`:
median(1:4)
median2(1:4)

median(c(1:3, 100, 1000))
median2(c(1:3, 100, 1000))

# With some `NA`s, the median can
# sometimes still be determined...
median2(c(0, 1, 1, 1, NA))
median2(c(0, 0, NA, 0, 0, NA, NA))

# ...unless there are too many `NA`s...
median2(c(0, 1, 1, 1, NA, NA))

# ...or too many unique values:
median2(c(0, 1, 2, 3, NA))
}
\author{
Lukas Jung, R Core Team
}
