% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-plot-col.R
\name{median_plot_col}
\alias{median_plot_col}
\title{Visualize rates of ignored \code{NA}s}
\usage{
median_plot_col(
  data,
  bar_alpha = 0.8,
  bar_color_na = "#F77774",
  bar_color_all = "#747DF7",
  ring_color = "black",
  ring_size = 8,
  show_ring = TRUE,
  show_legend = TRUE
)
}
\arguments{
\item{data}{Data frame returned by \code{\link[=median_table]{median_table()}}.}

\item{bar_alpha}{Numeric. Opacity of the bars. Default is \code{0.4}.}

\item{bar_color_na, bar_color_all}{Strings. Colors of the bars representing
the number of missing values that had to be ignored as a share of all
missing values (\verb{_na}) or of the entire sample (\verb{_all}).}

\item{ring_color}{String. Color of any "ring of certainty" half circle.
Default is \code{"black"}.}

\item{ring_size}{Numeric. Size of any "ring of certainty" half circle.
Default is \code{8}.}

\item{show_ring}{Logical. Should samples with a known median be marked by a
"ring of certainty" half circle? Default is \code{TRUE}.}

\item{show_legend}{Logical. Should a legend be displayed? Default is \code{TRUE}.
Note: there is no legend if there are no bars.}
}
\value{
A ggplot object.
}
\description{
\code{median_plot_col()} visualizes the results of
\code{\link[=median_table]{median_table()}}. It shows the rates of missing values that had to be
ignored to estimate the median of the remaining values.
}
\section{Visual guide (default)}{

\itemize{
\item Red bars show the share of missing values that had to be ignored as a share
of all missing values.
\item Blue bars show the same but as a share of \emph{all} values, missing or not.
They cover part of the blue bars; both types of bars start at zero.
\item The y-axis is fixed between 0 and 1 for a consistent display of
proportions.
\item Samples without any bar do not require ignoring any \code{NA}s, so the median is
known. They are also marked by a "ring of certainty", which is just a half
circle here.
}
}

\examples{
# Example data:
data <- median_table(
  list(
    c(0, 1, 1, 1, NA),
    c(1, 1, NA),
    c(1, 2, NA),
    c(0, 0, NA, 0, 0),
    c(1, 1, 1, 1, NA, NA),
    c(1, 1, 1, 1, NA, NA, NA),
    c(1, 1, 1, 1, NA, NA, NA, NA),
    iris$Sepal.Length,
    c(5.6, 5.7, 5.9, 6, 6.1, 6.3, 6.4, 6.6, 6.7, NA),
    c(6.1, 6.3, 5.9, 6, 6.1, 6.3, 6.4, 6.6, 6.7, NA, NA, NA, NA),
    c(7, 7, 7, 8, NA, NA)
  )
)

data

# See visual guide above
median_plot_col(data)
}
\seealso{
\itemize{
\item \code{\link[=median_plot_errorbar]{median_plot_errorbar()}} for an alternative visualization.
\item \code{\link[=median_table]{median_table()}} for the basis of these plots.
}
}
