% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-table.R
\name{median_table}
\alias{median_table}
\title{Tabulate median estimates with the certainty about them}
\usage{
median_table(x, even = c("mean", "low", "high"))
}
\arguments{
\item{x}{Vector or list of vectors. Each vector needs to be numeric or
similar. Note that data frames are lists, so \code{x} can be a data frame.}

\item{even}{Passed on to \code{\link[=median2]{median2()}}.}

\item{...}{Optional further arguments for \code{\link[=median2]{median2()}} methods. Not used in
its default method.}
}
\value{
Data frame with these columns:
\itemize{
\item \code{term}: the names of \code{x} elements. Only present if any are named.
\item \code{estimate}: the medians of \code{x} elements, ignoring as many \code{NA}s as
necessary.
\item \code{certainty}: \code{TRUE} if the corresponding estimate is certain to be the
true median, and \code{FALSE} if this is unclear due to missing values.
\item \code{min}, \code{max}: Lower and upper bounds of the median. Equal if \code{certainty}
is \code{TRUE} because in that case, the precise value is known.
\item \code{na_ignored}: the number of missing values that had to be ignored to
arrive at the estimate.
\item \code{na_total}: the total number of missing values.
\item \code{rate_ignored_na}: the proportion of missing values that had to be
ignored from among all missing values.
\item \code{sum_total}: the total number of values, missing or not.
\item \code{rate_ignored_sum}: the proportion of missing values that had to be
ignored from among all values, missing or not.
}
}
\description{
\code{median_table()} computes the sample median. Where the median is
unknown due to missing values, it only ignores as many of them as
necessary.

Estimates are presented along with information about whether they are known
to be the true median, how many missings had to be ignored, the rate of
ignored values, etc.

The function can also take a data frame (or another list) of numeric
vectors. It will then compute the median of each element.
}
\details{
The function deals with missing values (\code{NA}s) by first checking
whether they make the true median unknown. If they do, it removes one \code{NA},
then checks again; and so on until an estimate is found.

This strategy is based on \code{\link[=median2]{median2()}} and its \code{na.rm.amount} argument,
which represents a middle way between simply ignoring all \code{NA}s and not
even trying to compute an estimate. Instead, it only removes the minimum
number of \code{NA}s necessary, because some distributions have a known median
even if some of their values are missing. By keeping track of the removed
\code{NA}s, \code{median_table()} quantifies the uncertainty about its estimates.

The \code{min} and \code{max} columns are based on \code{\link[=median_range]{median_range()}}.
}
\examples{
median_table(c(5, 23, 5, NA, 5, NA))

# Use a list of numeric vectors:
my_list <- list(
  a = 1:15,
  b = c(1, 1, NA),
  c = c(4, 4, NA, NA, NA, NA),
  d = c(96, 24, 3, NA)
)

median_table(my_list)

# Data frames are allowed:
median_table(iris[1:4])
}
