% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-range.R
\name{median-possible}
\alias{median-possible}
\alias{median_possible_values}
\alias{median_range}
\alias{median_possible_values.default}
\alias{median_range.default}
\title{Possible medians}
\usage{
median_possible_values(x)

median_range(x)

\method{median_possible_values}{default}(x)

\method{median_range}{default}(x)
}
\arguments{
\item{x}{Numeric or similar. Vector to search for its possible medians.}
}
\value{
Vector of the same type as \code{x}. Its values are unique and sorted by
size. It always has length 2 when calling \code{median_range()}.

If the median can be determined (i.e., \code{median_na()} would return a
non-\code{NA} value), \code{median_possible_values()} returns just one value, and
\code{median_range()} returns two identical values.
}
\description{
These functions are helpful if \code{median_na()} returns \code{NA}:
\itemize{
\item \code{median_possible_values()} returns all values that might possibly be the
median, given the number of missing values.
\item \code{median_range()} returns the minimal and maximal possible median values.
}
}
\details{
Like \code{median_na()}, these functions are generic, so methods can be
defined for other classes. This documentation describes the default
methods.
}
