% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-na.R
\name{median2}
\alias{median2}
\alias{median2.default}
\title{Median with missing values analysis}
\usage{
median2(x, na.rm = FALSE, ...)

\method{median2}{default}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{Numeric or similar. Vector to search for its median.}

\item{na.rm}{Boolean. If set to \code{TRUE}, missing values are removed before
computation proceeds. Default is \code{FALSE}.}

\item{...}{Optional further arguments for methods. Not used in the default
method.}
}
\value{
Length-1 vector of the same type as \code{x}. The only exception occurs if
\code{x} is Boolean or integer and its length is even, in which case the return
value is double.

The output is \code{NA} (of the same type as \code{x}) if and only if the median
can't be determined because of missing values, or if there are no values.
}
\description{
Compute the sample median.

\code{median2()} works like the standard \code{\link[=median]{median()}} unless one or more values
are missing: \code{median()} always returns \code{NA} in this case, but \code{median2()}
checks if the median can be determined nevertheless.
}
\details{
\code{median2()} is a generic function, so new methods can be defined for
it. As with \code{\link[stats:median]{stats::median()}} from base R, the default method described
here should work for most classes for which a median is a reasonable
concept (e.g., "\code{\link{Date}}").

If a new method is necessary, please make sure it deals with missing values
like \code{median2.default()} does. See
\href{https://lhdjung.github.io/naidem/articles/algorithm.html}{\emph{Implementing
the algorithm}} for further details.
}
\examples{
# If no values are missing,
# it works like `median()`:
median(1:4)
median2(1:4)
median(c(1:3, 100, 1000))
median2(c(1:3, 100, 1000))

# With some `NA`s, the median can
# sometimes still be determined...
median2(c(0, 1, 1, 1, NA))
median2(c(0, 0, NA, 0, 0, NA, NA))

# ...unless there are too many `NA`s...
median2(c(0, 1, 1, 1, NA, NA))

# ...or too many unique values:
median2(c(0, 1, 2, 3, NA))
}
\author{
Lukas Jung, R Core Team
}
