% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-count-tolerable.R
\name{median_count_tolerable}
\alias{median_count_tolerable}
\title{How many \code{NA}s can be tolerated for a median estimate?}
\usage{
median_count_tolerable(x, needs_prep = TRUE)
}
\arguments{
\item{x}{A numeric vector or similar.}

\item{needs_prep}{Logical. Ignore unless the function is used as a helper.
See details.}
}
\value{
Integer (length 1). Never \code{NA}, never negative.
}
\description{
\code{median_count_tolerable()} returns the number of missing values
that can be preserved while determining the median. The point is to retain
as many data points as possible, instead of simply ignoring all \code{NA}s.

This is only based on the number of known values, not any \code{NA}s there might
be.

It is used within \code{\link[=median_table]{median_table()}} to determine how many missing values
need to be ignored.
}
\details{
With the default \code{needs_prep = TRUE}, missing values will be removed
from \code{x}, and \code{x} will be sorted. If this was already done elsewhere,
setting \code{needs_prep} to \code{FALSE} is more efficient. Proceed with caution as
this is not checked.
}
\examples{
# With two or fewer `NA`s, the median can only be `8`,
# so these `NA`s are tolerated:
median_count_tolerable(c(8, 8, 8, NA, NA))

# When adding a third `NA`, the median will be unknown.
# Compare using naidem's correct median function:
median2(c(8, 8, 8, NA, NA))
median2(c(8, 8, 8, NA, NA, NA))

# No `NA`s are tolerable here because
# a single one could change the median:
median_count_tolerable(c(8, 9, 9, NA, NA, NA))

# Here too, the median depends on the value behind `NA`,
# so the `NA` cannot be tolerated:
median_count_tolerable(c(8, 9, NA))
}
