% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-is-determinate.R
\name{median_is_determinate}
\alias{median_is_determinate}
\title{Can the median be determined?}
\usage{
median_is_determinate(x, na.rm.amount = 0)
}
\arguments{
\item{x}{Numeric or similar. Vector to search for its median.}

\item{na.rm.amount}{Numeric. Alternative to \code{na.rm} that only removes a
specified number of missing values. Default is \code{0}.}
}
\value{
Logical (length 1).
}
\description{
\code{median_is_determinate()} tests whether a vector's median can be determined
despite missing values. It is based on \code{\link[=median2]{median2()}}.
}
\examples{
# The median is 1, no matter which number
# `NA` represents:
median_is_determinate(c(1, 1, NA))

# Here, the median may be 1 or 2, depending
# on the value behind `NA`; so the median
# cannot be determined:
median_is_determinate(c(1, 2, NA))
}
