% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-plot.R
\name{median_plot}
\alias{median_plot}
\title{Visualize median estimate certainty}
\usage{
median_plot(data, line_width = 0.35)
}
\arguments{
\item{data}{Data frame returned by \code{\link[=median_table]{median_table()}}.}

\item{line_width}{Numeric (length 1). Width of the horizontal lines. Default
is \code{0.35}.}
}
\value{
A ggplot object.
}
\description{
Use \code{median_plot()} to visualize the results of
\code{\link[=median_table]{median_table()}}.
}
