% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-plot.R
\name{median_plot}
\alias{median_plot}
\title{Visualize median estimate uncertainty}
\usage{
median_plot(data, point_size = 2, line_width = 0.5, bar_width = 0.9)
}
\arguments{
\item{data}{Data frame returned by \code{\link[=median_table]{median_table()}}.}

\item{point_size}{Numeric. Size of the median estimate points. Default is
\code{2}.}

\item{line_width}{Numeric (length 1). Width of the error bar lines. Default
is \code{0.5}.}

\item{bar_width}{Numeric (length 1). Extension of the horizontal bars.
Default is \code{0.9}.}
}
\value{
A ggplot object.
}
\description{
\code{median_plot()} visualizes the results of \code{\link[=median_table]{median_table()}}. It
shows the \code{min} and \code{max} median bounds using error bars. Median estimates
are displayed as points.
}
\section{Visual guide}{

\itemize{
\item Points are medians of the non-\code{NA} values.
\item Points that are known to be true medians have a "ring of certainty"
around them.
\item Error bars display lower and upper bounds of the true median, reflecting
any uncertainty created by missing values. The median is known to fall
between the two error bars, even if its exact value is unknown.
\item If no bounds can be found for a sample because of too many missing
values, the error bars span the height of the plot, and the point is a
hexagram. The median is particularly uncertain in this case because it
cannot even be confined to a range. See \code{\link[=median_range]{median_range()}}.
}
}

\examples{
# Example data:
data <- median_table(
  list(
    c(0, 1, 1, 1, NA),
    c(1, 1, NA),
    c(1, 2, NA),
    c(0, 0, NA, 0, 0),
    c(1, 1, 1, 1, NA, NA),
    c(1, 1, 1, 1, NA, NA, NA),
    c(1, 1, 1, 1, NA, NA, NA, NA),
    iris$Sepal.Length,
    c(5.6, 5.7, 5.9, 6, 6.1, 6.3, 6.4, 6.6, 6.7, NA),
    c(6.1, 6.3, 5.9, 6, 6.1, 6.3, 6.4, 6.6, 6.7, NA, NA, NA, NA),
    c(7, 7, 7, 8, NA, NA)
  )
)

data

# -- Some medians are known: "ring of certainty" and no
# error bar expansion.
# -- Some medians are unknown but confined to a range.
# -- One median doesn't even have a range, its error bars
# extend into infinity, and its point is a hexagram.
median_plot(data)
}
